Use SSH instead of the serial port for local access
This is a tough decision

Originally, I thought that making local console access should work
without an SSH server. And I still think so, and, actually, this access
remains - via the command.sock
However, there is no concept of a session for the serial console, at
least not something out of the box. Not having a session means no
determinism. You'll either be prompted to enter the user name on login,
either not. This makes scripting via serial unnecessarily complicatd.

On the other hand, adding SSHD to the base image, and putting an SSH
key to the VM's authorized_keys, and forwarding the port 22 is relatively
trivial.

I am not totally happy about this, and will look into making the serial
console the default method in the future.

As a side note, making the serial console work in interactive mode with
the shell is not straightforward.
nc and socat over the socket don't interpret escape sequences.
pty suddenly stopped working
telnet is horrible with its protocol negotiation and the fallback
"legacy line protocol". Switching protocols doesn't always work as
expected. I could never make it work without it littering the output
with some garbage (like a question mark).

To make things even more complicated, there can be only one serial
console, two or more did not work for me.
It is also shared, so if you're in an interactive console, and also
want to run a script, it will happen in the same stream.

The only combination of protocol, chardev, and client that I could make
to work was the tio CLI client connecting to a pty  device.
socat works perfectly as a proxy to a UNIX socket instead of a pty device
with QEMU, effectively allowing one connection over pty, and a completely
separate over socket.